Handle unavailable courses; new update method
Bugs fixed:
#Now able to show message about a course which is no longer available
for the current semester, instead of loading the data forever.
#Course list in major plan never shows two courses with the same
continuity id but different codes.
#Removed "Start over" button when the user is modifying the plan.
-----------------------
Now when loading a plan/schedule, it will check if any of the chosen
sections is unavailable. If so, a notification will shows up and the
color will also change to indicate the section is unavailable.

Now it will use two collections to update the database. During an
update, while one of them is being modified, the other collection has
the complete new data of previous update.