Improved performance for course list
Reduced the amount of search when checking whether the course has been
added or getting the color of a future course.

Added a new feature so that now when the user added a course to the
card, the course will hide in the course list unless the user chooses
not to automatically hide added courses.