Bug fixed
#Changed the order of inserting new terms. They are now inserted after
the update is completely done. And client will refresh once the new
terms are pushed, so that if the user is creating/modifying/viewing a
plan, it can prevent most of the error and data loss as well as making
sure the user gets the newest data in real time.